Handled build errors
