+++ .eslintrc.json
@@ -1,4 +1,12 @@
 {
-  "extends": ["next/core-web-vitals", "next/typescript"],
-  "@typescript-eslint/no-explicit-any": "off"
+  "extends": [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/recommended"
+  ],
+  "parser": "@typescript-eslint/parser",
+  "plugins": ["@typescript-eslint"],
+  "rules": {
+    "@typescript-eslint/no-explicit-any": "off",  // Turn off the rule
+    // You can add other rules here as needed
+  }
 }
