+++ public/sw.js
@@ -1,93 +1,1 @@
-/**
- * Copyright 2018 Google Inc. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// If the loader is already loaded, just stop.
-if (!self.define) {
-  let registry = {};
-
-  // Used for `eval` and `importScripts` where we can't get script URL by other means.
-  // In both cases, it's safe to use a global var because those functions are synchronous.
-  let nextDefineUri;
-
-  const singleRequire = (uri, parentUri) => {
-    uri = new URL(uri + ".js", parentUri).href;
-    return registry[uri] || (
-      
-        new Promise(resolve => {
-          if ("document" in self) {
-            const script = document.createElement("script");
-            script.src = uri;
-            script.onload = resolve;
-            document.head.appendChild(script);
-          } else {
-            nextDefineUri = uri;
-            importScripts(uri);
-            resolve();
-          }
-        })
-      
-      .then(() => {
-        let promise = registry[uri];
-        if (!promise) {
-          throw new Error(`Module ${uri} didn’t register its module`);
-        }
-        return promise;
-      })
-    );
-  };
-
-  self.define = (depsNames, factory) => {
-    const uri = nextDefineUri || ("document" in self ? document.currentScript.src : "") || location.href;
-    if (registry[uri]) {
-      // Module is already loading or loaded.
-      return;
-    }
-    let exports = {};
-    const require = depUri => singleRequire(depUri, uri);
-    const specialDeps = {
-      module: { uri },
-      exports,
-      require
-    };
-    registry[uri] = Promise.all(depsNames.map(
-      depName => specialDeps[depName] || require(depName)
-    )).then(deps => {
-      factory(...deps);
-      return exports;
-    });
-  };
-}
-define(['./workbox-7144475a'], (function (workbox) { 'use strict';
-
-  importScripts();
-  self.skipWaiting();
-  workbox.clientsClaim();
-  workbox.registerRoute("/", new workbox.NetworkFirst({
-    "cacheName": "start-url",
-    plugins: [{
-      cacheWillUpdate: async ({
-        response: e
-      }) => e && "opaqueredirect" === e.type ? new Response(e.body, {
-        status: 200,
-        statusText: "OK",
-        headers: e.headers
-      }) : e
-    }]
-  }), 'GET');
-  workbox.registerRoute(/.*/i, new workbox.NetworkOnly({
-    "cacheName": "dev",
-    plugins: []
-  }), 'GET');
-
-}));
-//# sourceMappingURL=sw.js.map
+if(!self.define){let e,s={};const c=(c,t)=>(c=new URL(c+".js",t).href,s[c]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=c,e.onload=s,document.head.appendChild(e)}else e=c,importScripts(c),s()})).then((()=>{let e=s[c];if(!e)throw new Error(`Module ${c} didn’t register its module`);return e})));self.define=(t,a)=>{const n=e||("document"in self?document.currentScript.src:"")||location.href;if(s[n])return;let i={};const r=e=>c(e,n),o={module:{uri:n},exports:i,require:r};s[n]=Promise.all(t.map((e=>o[e]||r(e)))).then((e=>(a(...e),i)))}}define(["./workbox-f1770938"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/4c91R_Pm5Xc--hSZtD8X2/_buildManifest.js",revision:"a0ae24e7f29dd3809ab75b5dd91a79dc"},{url:"/_next/static/4c91R_Pm5Xc--hSZtD8X2/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/138-9aeee29c14dc71f0.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/190-4e9863e6030553c4.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/349-42647c649cb666a4.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/52ab8b6c-409f569fdab512db.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/53c13509-1aac23ceefca6f8e.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/595-6224dc4b48e23788.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/5e22fd23-1ac0e6f2f3c39f5a.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/636-232f5e347d8b1ab6.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/795d4814-15a1a511d05b573b.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/978-a05d29a324572e67.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/998-fba0e81e0389f582.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/9c4e2130-c717f6abdacf9f0c.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/aaea2bcf-c96f4eca0af8e8a1.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/app/_not-found/page-2441a29fc304d637.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/app/billing/page-2dd96406af1c42f8.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/app/components/custom-table/page-93add7a147863a84.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/app/dashboard/page-cdc472da694efa8e.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/app/instances/create-instance/page-70d26ecdaf6bb355.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/app/instances/page-4ec8e16e3e294f08.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/app/layout-e598e3ad3cbafcb6.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/app/login/page-f91ac771a0253648.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/app/page-c44a2526ea60d348.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/app/register/page-d65b01afa19f31ba.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/eec3d76d-6e35b433670c16d8.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/fc2f6fa8-84e407f3d06e92ef.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/fca4dd8b-603de03a6fb72ce3.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/fd9d1056-3a9d801b0e831361.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/framework-00a8ba1a63cfdc9e.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/main-app-0cb1bc39dbbd2d02.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/main-ef7deec28afb5300.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/pages/_app-037b5d058bd9a820.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/pages/_error-6ae619510b1539d6.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-1101b4dba4de7ab4.js",revision:"4c91R_Pm5Xc--hSZtD8X2"},{url:"/_next/static/css/3f0ce0cb2c5e4fdf.css",revision:"3f0ce0cb2c5e4fdf"},{url:"/_next/static/css/70edc9936b8e3f66.css",revision:"70edc9936b8e3f66"},{url:"/_next/static/css/bd02915611f90898.css",revision:"bd02915611f90898"},{url:"/_next/static/media/github.0ddaff9f.png",revision:"aeb56c018fc0aa317bc9b97cf0b17964"},{url:"/_next/static/media/google.aeee8a20.png",revision:"e91b3a96a67a4ef0a7f60d26a45f8277"},{url:"/cloud_servers.png",revision:"a22b831d7909a5e65ebb2b181d72cb0d"},{url:"/github.png",revision:"aeb56c018fc0aa317bc9b97cf0b17964"},{url:"/google.png",revision:"e91b3a96a67a4ef0a7f60d26a45f8277"}],{ignoreURLParametersMatching:[/^utm_/,/^fbclid$/]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({response:e})=>e&&"opaqueredirect"===e.type?new Response(e.body,{status:200,statusText:"OK",headers:e.headers}):e}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:2592e3})]}),"GET"),e.registerRoute(/\/_next\/static.+\.js$/i,new e.CacheFirst({cacheName:"next-static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4|webm)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:48,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({sameOrigin:e,url:{pathname:s}})=>!(!e||s.startsWith("/api/auth/callback")||!s.startsWith("/api/"))),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({request:e,url:{pathname:s},sameOrigin:c})=>"1"===e.headers.get("RSC")&&"1"===e.headers.get("Next-Router-Prefetch")&&c&&!s.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages-rsc-prefetch",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({request:e,url:{pathname:s},sameOrigin:c})=>"1"===e.headers.get("RSC")&&c&&!s.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages-rsc",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:{pathname:e},sameOrigin:s})=>s&&!e.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({sameOrigin:e})=>!e),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
