+++ src/app/login/page.tsx
@@ -11,7 +11,6 @@ import {
 // import { getServerSession } from "next-auth";
 // import { redirect } from "next/navigation";
 import { CredentialsForm } from "@/app/components/credentialsForm";
-import { authOptions } from "@/app/lib/auth";
 import { useRouter } from "next/navigation";
 import LoginForm from "../components/loginForm";
 import Link from "next/link";
@@ -20,7 +19,6 @@ import Link from "next/link";
 export default function SignInPage() {
   const baseUrl = process.env.NEXT_PUBLIC_VMMS_BACKEND_URL;
   const router = useRouter();
-  // const session = await getServerSession(authOptions);
 
   // console.log("Session: ", session);
